+++ app/Http/Controllers/Api/SchoolController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\SchoolResource;
 use App\Models\School;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 
 class SchoolController extends Controller
@@ -119,6 +120,46 @@ class SchoolController extends Controller
     public function destroy(School $id)
     {
         // RegistrationForm, StudentRegistration, Testimony, User
+        if (DB::table('users')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data User.!'
+            ], 422);
+        }
+        if (DB::table('achievements')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data Penghargaan.!'
+            ], 422);
+        }
+        if (DB::table('facilities')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data Fasilitas.!'
+            ], 422);
+        }
+        if (DB::table('extracurriculars')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data Ekstrakurikuler.!'
+            ], 422);
+        }
+        if (DB::table('galleries')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data Galeri.!'
+            ], 422);
+        }
+        if (DB::table('registration_forms')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data Form Registrasi.!'
+            ], 422);
+        }
+        if (DB::table('teachers')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data Guru.!'
+            ], 422);
+        }
+        if (DB::table('testimonies')->where('school_id', $id->id)->count()) {
+            return response()->json([
+                'msg' => 'Data Sekolah ini masih menjadi referensi di data Testimoni.!'
+            ], 422);
+        }
         $id->delete();
         return response()->json([
             'data'  => $id,
